requirements.txt
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.2.1
dnspython==2.8.0
Flask==3.1.2
idna==3.11
imutils==0.5.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.6
opencv-python==4.12.0.88
paho-mqtt==2.1.0
pandas==3.0.0
pymongo==4.16.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.5
six==1.17.0
urllib3==2.5.0
Werkzeug==3.1.3
